Factor out TB validation test function
