+++ Makefile
@@ -62,7 +62,7 @@ release: export BINARY_NAME = uci
 release: export MODULE_CHOICES = uci
 test: export MODULE_NAME = unittest
 test: export BINARY_NAME = unittest
-test: export MODULE_CHOICES = {unittest,bookgen,gamefile}
+test: export MODULE_CHOICES = {unittest,bookgen,gamefile,tbvalidate}
 perft: export MODULE_NAME = perft
 perft: export BINARY_NAME = perft
 perft: export MODULE_CHOICES = perft
