+++ src/syzygy_probe.hpp
@@ -1,3 +1,5 @@
+# pragma once
+
 # include <string>
 # include <tuple>
 # include <list>
