+++ src/tbvalidate/test_func.hpp
@@ -1,0 +1,4 @@
+# include <string>
+# include "../syzygy_probe.hpp"
+
+void test_syzygy_probing_vs_known_fens(Tablebase &tb, const std::string fen_csv_file);
